% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/{{cookiecutter.package_name}}-package.R
\docType{package}
\name{{ "{" }}{{cookiecutter.package_name}}-package{{ "}" }}
\alias{{ "{" }}{{cookiecutter.package_name}}{{ "}" }}
\alias{{ "{" }}{{cookiecutter.package_name}}-package{{ "}" }}
\title{{ "{" }}{{cookiecutter.package_title}}{{ "}" }}
\description{
{{cookiecutter.package_description}}
}
\author{
\strong{Maintainer}: {{cookiecutter.author}} \email{{ "{" }}{{cookiecutter.email}}{{ "}" }} (\href{https://orcid.org/{{cookiecutter.orcid}}}{ORCID})

}
\keyword{internal}
